SeniorCapstoneProject2020-main/requirements.txt
discord.py==1.6.0
aiohttp==3.7.3
async-timeout==3.0.1
chardet==3.0.4
multidict==5.1.0
typing-extensions==3.7.4.3
yarl==1.6.3
idna==2.8
discord-timers==0.1.0
better-profanity==0.7.0
validators==0.18.2
decorator==5.0.7
six==1.15.0